+++ db/migrate/20190508214541_create_projects.rb
@@ -1,0 +1,9 @@
+class CreateProjects < ActiveRecord::Migration[6.0]
+  def change
+    create_table :projects do |t|
+      t.string :name
+      t.string :access_token
+      t.timestamps
+    end
+  end
+end