+++ db/migrate/20190508215106_create_report_rows.rb
@@ -1,0 +1,15 @@
+class CreateReportRows < ActiveRecord::Migration[6.0]
+  def change
+    create_table :report_rows do |t|
+      t.bigint :report_id
+      t.string :label
+      t.text :expression
+      t.integer :position
+      t.jsonb :result
+      t.timestamps
+
+      t.index :report_id
+      t.index :position
+    end
+  end
+end